Enhance colors handing in `svg-icon'
Some color names known to Emacs are not known to svg, for example
"red3".  This patch makes it possible to specify Emacs colors as
FG-COLOR and BG-COLOR arguments to `svg-icon', they are converted to
svg colors.

Also, makes it possible to specify FG-COLOR and BG-COLOR as ordinary
Emacs faces.  For FG-COLOR `:foreground' attribute is taken as color,
for BG-COLOR `:background' attribute is taken as color.  If face has
no such attribute, then fallback to defaults.